refactor storage a módulos aparte
