+++ infra/apps/catalog/helpers/file_name_for_format.py
@@ -1,0 +1,7 @@
+def file_name_for_format(catalog):
+    names = {
+        'json': 'data',
+        'xlsx': 'catalog'
+    }
+    file_name = names[catalog.format]
+    return file_name
